+++ modules/m_admin.c
@@ -54,7 +54,7 @@ _admin_process_request(struct Client *source)
  *      - parv[1] = nickname/servername
  */
 static void
-m_admin(struct Client *source, int parc, char *parv[])
+m_admin(struct Client *source, size_t parc, char *parv[])
 {
   static uintmax_t last_used = 0;
 
@@ -85,7 +85,7 @@ m_admin(struct Client *source, int parc, char *parv[])
  *      - parv[1] = nickname/servername
  */
 static void
-ms_admin(struct Client *source, int parc, char *parv[])
+ms_admin(struct Client *source, size_t parc, char *parv[])
 {
   if (server_route_command(source, ":%s ADMIN :%s", 1, parv)->result != SERVER_ROUTE_ISME)
     return;
