+++ modules/m_kline.c
@@ -178,7 +178,7 @@ kline_handle(struct Client *source, const struct aline_ctx *aline)
 }
 
 static void
-mo_kline(struct Client *source, int parc, char *parv[])
+mo_kline(struct Client *source, size_t parc, char *parv[])
 {
   if (!client_has_oper_flag(source, OPER_FLAG_KLINE))
   {
@@ -222,7 +222,7 @@ mo_kline(struct Client *source, int parc, char *parv[])
  *      - parv[5] = reason
  */
 static void
-ms_kline(struct Client *source, int parc, char *parv[])
+ms_kline(struct Client *source, size_t parc, char *parv[])
 {
   uintmax_t duration;
   if (io_parse_uintmax(parv[2], &duration) != IO_PARSE_OK)
