+++ modules/m_metadata.c
@@ -29,7 +29,7 @@
  *      - parv[0] = command
  */
 static void
-ms_metadata(struct Client *source, int parc, char *parv[])
+ms_metadata(struct Client *source, size_t parc, char *parv[])
 {
   if (!client_is_service(source) && !client_is_server(source))
     return;
