+++ modules/m_svskill.c
@@ -32,7 +32,7 @@
  *      - parv[3] = kill message
  */
 static void
-ms_svskill(struct Client *source, int parc, char *parv[])
+ms_svskill(struct Client *source, size_t parc, char *parv[])
 {
   if (!client_is_service(source) && !client_is_server(source))
     return;
