+++ modules/m_svstag.c
@@ -28,7 +28,7 @@
  *                 pointers.
  */
 static void
-ms_svstag(struct Client *source, int parc, char *parv[])
+ms_svstag(struct Client *source, size_t parc, char *parv[])
 {
   if (!client_is_service(source) && !client_is_server(source))
     return;
